pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "cubingbot"
version = "0.1.0"
[tool.poetry.dependencies]
gspread = "^3.6"
oauth2client = "^4.1"
python = "^3.8"
flask = "^1.1"
google-api-python-client = "^1.9"
[tool.poetry.dev-dependencies]
